[optimization] [Service] Optimization ProcessService and add ProcessService interface (#9370)
